+++ CHANGES.md
@@ -2,6 +2,8 @@
 
 ## [Unreleased](https://github.com/Amebis/eduVPN/compare/4.1.1...HEAD)
 
+- eduvpn-common updated to discovery-improvements ab95a24dcf2404899543638149b0a44a5d2720a3
+
 
 ## [4.1.1](https://github.com/Amebis/eduVPN/compare/4.1...4.1.1) (2024-07-01)